Move styles to sass
