+++ components/About/style.scss
@@ -31,4 +31,11 @@
     color: #813d8f;
     margin: 15px auto;
   }
+  > hr {
+    border-style: solid;
+    border-width: 0.5px;
+    border-color: #813d8f;
+    margin-top: 15px;
+    margin-bottom: 30px;
+  }
 }
