+++ components/Footer/style.scss
@@ -1,0 +1,58 @@
+.footer {
+  background-color: #18263a;
+  padding: 30px;
+
+  .warsawex_logo {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  .symbol {
+    height: 102px;
+    width: 73px;
+  }
+  .logo {
+    height: 25px;
+    width: auto;
+    fill: white;
+    margin-top: 22px;
+  }
+  .media {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    margin-top: 42px;
+    margin-bottom: 32px;
+  }
+  .media img {
+    margin: 0px 8px;
+    height: 19px;
+    width: auto;
+  }
+  .menu {
+    color: white;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  h2 {
+    font-size: 15px;
+    font-weight: 500;
+    margin: 0px 0px 13px 0px;
+  }
+  .contact {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  .contact > h2 {
+    font-weight: 600;
+    margin-top: 20px;
+    margin-bottom: 20px;
+  }
+  .contact > h1 {
+    font-size: 18px;
+    font-weight: 500;
+    margin-top: 0px;
+  }
+}
