+++ components/Founders/style.scss
@@ -1,0 +1,33 @@
+.founders {
+  padding: 30px;
+  text-align: center;
+  > h1 {
+    font-size: 33px;
+    font-weight: 300;
+    color: #813d8f;
+    margin: 15px auto;
+  }
+  hr {
+    border-style: solid;
+    border-width: 0.5px;
+    border-color: #813d8f;
+    margin-top: 15px;
+    margin-bottom: 30px;
+  }
+  > img {
+    height: 174px;
+    width: 174px;
+  }
+  > h2 {
+    font-size: 20px;
+    font-weight: 700;
+    color: #707070;
+    margin-bottom: 10px;
+  }
+  > h3 {
+    font-size: 20px;
+    font-weight: 300;
+    color: #707070;
+    margin: 0px 0px 50px 0px;
+  }
+}
