+++ components/Layout/style.scss
@@ -1,0 +1,5 @@
+body {
+  font-family: "Montserrat", sans-serif;
+  margin: 0;
+  padding: 0;
+}
