+++ components/MenuToggle/index.js
@@ -1,0 +1,11 @@
+import style from "./style.scss";
+
+const MenuToggle = props => (
+  <div className={style.menu} onClick={props.onClick}>
+    <div />
+    <div />
+    <div />
+  </div>
+);
+
+export default MenuToggle;
