+++ components/MenuToggle/style.scss
@@ -1,0 +1,6 @@
+.menu > div {
+  width: 40px;
+  height: 3px;
+  background-color: #132233;
+  margin-bottom: 5px;
+}
