+++ components/NextMeetup/style.scss
@@ -1,0 +1,97 @@
+.next {
+  text-align: center;
+  padding: 30px;
+  .num {
+    display: inline-flex;
+    flex-wrap: nowrap;
+    padding: 0px;
+    margin: 0px;
+    line-height: 140px;
+    > :first-child {
+      font-size: 125px;
+      font-weight: 700;
+      margin-right: 18px;
+    }
+    > :last-child {
+      font-size: 180px;
+      font-weight: 900;
+      margin: 0px;
+    }
+  }
+  > .num,
+  h2 {
+    color: #813d8f;
+    opacity: 0.84;
+  }
+  h2 {
+    font-size: 54px;
+    font-weight: 600;
+    margin-top: 0px;
+  }
+  .date {
+    font-size: 27px;
+    font-weight: 600;
+    color: #132233;
+    margin-top: 28px;
+  }
+  .location {
+    margin-top: 24px;
+    display: inline-flex;
+    align-items: center;
+  }
+  .location > p {
+    color: #707070;
+    font-size: 18px;
+    font-weight: 500;
+    margin-left: 15px;
+  }
+  .agenda > h1 {
+    font-size: 40px;
+    font-weight: 300;
+    color: #132233;
+    margin-bottom: 0px;
+  }
+  hr {
+    border-style: solid;
+    border-width: 0.5px;
+    border-color: #813d8f;
+    margin-top: 20px;
+    margin-bottom: 35px;
+  }
+  .talks > h2 {
+    font-size: 18px;
+    font-weight: 700;
+    color: #132233;
+    opacity: 1;
+    margin-bottom: 25px;
+  }
+  .talks > h3,
+  p {
+    color: #707070;
+  }
+  .talks > h3 {
+    font-size: 16px;
+    margin-top: 0px;
+    margin-bottom: 10px;
+  }
+  .talks > h2,
+  p {
+    font-size: 18px;
+  }
+  .talks p {
+    color: #000000;
+    margin: 30px 27px 0px 27px;
+  }
+  .level {
+    display: inline-flex;
+    margin-top: 15px;
+  }
+  .level > h3 {
+    color: #132233;
+    font-size: 16px;
+    margin-right: 10px;
+  }
+  .level > img {
+    margin: 10px 3px 70px 0px;
+  }
+}
