+++ components/SeeUs/index.js
@@ -1,0 +1,10 @@
+import style from "./style.scss";
+
+const SeeUs = () => (
+  <div className={style.see_us}>
+    See us on the next Meetup!
+    <button className={style.join}>JOIN THE COMMUNITY</button>
+  </div>
+);
+
+export default SeeUs;
