--- components/Sponsors.js
@@ -1,28 +1,0 @@
-const Sponsors = () => (
-  <div>
-    <h1>Our Patrons</h1>
-    <hr />
-    <style jsx>
-      {`
-        div {
-          text-align: center;
-          padding: 30px;
-        }
-        h1 {
-          font-size: 40px;
-          font-weight: 300;
-          color: #132233;
-          margin-bottom: 0px;
-        }
-        hr {
-          border-style: solid;
-          border-width: 0.5px;
-          border-color: #813d8f;
-          margin-top: 20px;
-          margin-bottom: 35px;
-      `}
-    </style>
-  </div>
-);
-
-export default Sponsors;
