+++ components/Sponsors/index.js
@@ -1,0 +1,10 @@
+import style from "./style.scss";
+
+const Sponsors = () => (
+  <div className={style.patrons}>
+    <h1>Our Patrons</h1>
+    <hr />
+  </div>
+);
+
+export default Sponsors;
