+++ components/Sponsors/style.scss
@@ -1,0 +1,17 @@
+.patrons {
+  text-align: center;
+  padding: 30px;
+  > h1 {
+    font-size: 40px;
+    font-weight: 300;
+    color: #132233;
+    margin-bottom: 0px;
+  }
+  hr {
+    border-style: solid;
+    border-width: 0.5px;
+    border-color: #813d8f;
+    margin-top: 20px;
+    margin-bottom: 35px;
+  }
+}
